requirements.txt
flask
colorama
requests
dnspython
python-whois
beautifulsoup4
python-dotenv
pillow
exifread
lxml
tldextract
phonenumbers
validators
aiohttp
